RL_LLM/requirements.txt
importlib-metadata<5.0
spacy==3.0.6
bert-score==0.3.11
BLEURT @ git+https://github.com/google-research/bleurt.git@c6f2375c7c178e1480840cf27cb9e2af851394f9
datasets==2.5.1
gem-metrics @ git+https://github.com/GEM-benchmark/GEM-metrics.git
gym==0.21.0
jsonlines==3.0.0
nltk==3.7
pandas==1.3.5
rich==12.0.0
stable-baselines3==1.5.1a5
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
huggingface-hub==0.11.1
wandb==0.12.15
jsonlines==3.0.0
rouge_score==0.0.4
sacrebleu==2.2.0
py-rouge==1.1
absl-py
six
progressbar==2.5
openai==0.27.2
evaluate==0.4.0
yake==0.4.8
keybert==0.7.0
pytextrank==3.2.4
string2string==0.0.145